final commit #1

+++ myApp/app/Http/Controllers/AdminController.php
@@ -16,18 +16,40 @@ use App\Models\Team_Calendar;
 use App\Models\Admin;
 use App\Models\Super_Admin;
 use App\Models\Account_Manager;
+use Illuminate\Support\Facades\Crypt;
 
 class AdminController extends Controller
 {
+    public function userChecker(Request $request){
+        if (session()->get('key') != 'Admin'){
+            $request->session()->flash('error','Cannot access this page' );
+            return false;
+        }
+        else{
+            return true;
+        }
+
+    }
+
+
     public function viewSlots(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         return view('admin' , ['slots'=> Brief::where('status' , 'Approved')->get()]);
     }
 
-    public function viewUsers(Request $request){
-        return view('admin/manageusers' , ['users'=> User::all()]);
+    public function viewUsersandManagers(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
+        return view('admin/manageusers' , ['users'=> User::all(),'managers' => Account_Manager::all()]);
     }
 
     public function viewUserData(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         $username = User::where('user_id',$request->userID)->value('username'); 
         $userID = $request->userID;
       
@@ -36,8 +58,59 @@ class AdminController extends Controller
         'feedbacks' => Feedback::where('user_id' , $userID)->get() , 'problems' => Problem::where('user_id' , $userID)->get() ,
         'senders'=>Financial_History::where('sender',$username)->get(),'receivers'=>Financial_History::where('receiver',$username)->get() ]);
     }
+    public function addUser(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
+        $newAccount = new System_Accounts;
+        $newUser = new User;
+
+
+        $newAccount->username = $request->username;
+        $newAccount->password = Crypt::encrypt($request->password);
+        $newAccount->email = Crypt::encrypt($request->email);
+        $newAccount->type = 'User';
+        $newAccount->timestamps = false;
+        $newAccount->save();
+
+        $newUser->username = $request->username;
+        $newUser->password = Crypt::encrypt($request->password);
+        $newUser->email = Crypt::encrypt($request->email);
+        $newUser->timestamps = false;
+        $newUser->save();
+
+        $benefit1 = new Benefit;
+        $benefit1->name = "Smart Gym Membership";
+        $benefit1->description = "Free 2 month Smart Gym Membership in all of their branches";
+        $benefit1->redeemed = 0;
+        $benefit1->user_id = User::where('username' , $request->username)->value('user_id');
+        $benefit1->timestamps = false;
+        $benefit1->save();
+
+        $benefit2 = new Benefit;
+        $benefit2->name = "Free L'oreal Product";
+        $benefit2->description = "Your choice of either face cleanser or moisturizer from L'oreal";
+        $benefit2->redeemed = 0;
+        $benefit2->user_id = User::where('username' , $request->username)->value('user_id');
+        $benefit2->timestamps = false;
+        $benefit2->save();
+
+        $benefit3 = new Benefit;
+        $benefit3->name = "Free Netflix Subscription";
+        $benefit3->description = "Free 2 month Netflix Membership";
+        $benefit3->redeemed = 0;
+        $benefit3->user_id = User::where('username' , $request->username)->value('user_id');
+        $benefit3->timestamps = false;
+        $benefit3->save();
+
+
+        return view('admin/manageusers' , ['users'=> User::all(),'managers' => Account_Manager::all()]);
+    }
 
     public function deleteUser(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         //TO DELETE A USER YOU NEED TO DELETE: their benefits, briefs, feedback, problem, FINANCIAL HISTORY AND MONEY TRANSACTION
         // THEN YOU NEED TO DELETE THE USER ITSELF AND THEN YOU CAN DELETE THE SYSTEM ACCOUNT
         
@@ -111,28 +184,40 @@ class AdminController extends Controller
             //throw $th;
         }
         
-        return view('admin/manageusers' , ['users'=> User::all()]);
+        return view('admin/manageusers' , ['users'=> User::all(),'managers' => Account_Manager::all()]);
     }
 
     public function assignManager(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         $user = User::where('user_id' , $request->userID)->first();
         $user->manager = $request->email;
         $user->timestamps =false;
         $user->save();
-        return view('admin/manageusers' , ['users'=> User::all()]);
+        return view('admin/manageusers' , ['users'=> User::all(),'managers' => Account_Manager::all()]);
     }
 
     public function viewProjects(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         return view('admin/manageprojectsandservices' , ['projects'=> Projects_and_Services::all()]);
 
     }
 
     public function deleteProject (Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         Projects_and_Services::where('prod_id' , $request->projID)->delete();
         return view('admin/manageprojectsandservices' , ['projects'=> Projects_and_Services::all()]);
     }
 
     public function addProject (Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         $project = new Projects_and_Services;
         $project->name = $request->name;
         $project->description = $request->description;
@@ -142,6 +227,9 @@ class AdminController extends Controller
     }
 
     public function editProject (Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         $project = Projects_and_Services::where('prod_id' , $request->projID)->first();
         $project->name = $request->name;
         $project->description = $request->description;
@@ -152,10 +240,16 @@ class AdminController extends Controller
     }
 
     public function viewBriefs(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         return view('admin/managebriefs' , ['briefs'=> Brief::all()]);
     }
 
     public function approveBrief(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         $brief = Brief::where('brief_id' , $request->briefID)->first();
         $brief->status = 'Approved';
         $brief->timestamps = false;
@@ -166,6 +260,9 @@ class AdminController extends Controller
     }
 
     public function denyBrief(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         $brief = Brief::where('brief_id' , $request->briefID)->first();
         $brief->status = 'Denied';
         $brief->timestamps = false;
@@ -175,10 +272,16 @@ class AdminController extends Controller
     }
 
     public function viewManagers(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         return view('admin/manageaccountmanagers' , ['managers'=> Account_Manager::all()]);
     }
 
     public function addManager(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         $manager = new Account_Manager;
         $manager->email = $request->email;
         $manager->name = $request->name;
@@ -190,6 +293,9 @@ class AdminController extends Controller
     }
 
     public function deleteManager(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         //REMOVE THEM FROM BEING ASSIGNED TO ANY USER FIRST
         $users = User::where('manager' , $request->email)->get();
         foreach ($users as $user) {
@@ -203,10 +309,16 @@ class AdminController extends Controller
     }
 
     public function viewFeedback(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         return view('admin/viewfeedback' , ['feedbacks'=> Feedback::all()]);
     }
 
     public function viewProblems(Request $request){
+        if ($this->userChecker($request) == false ){
+            return view('home');
+        }
         return view('admin/viewproblems' , ['problems'=> Problem::all()]);
     }
 }
